cube_project/requirements.txt
open3d>=0.19.0
numpy
scipy
pyyaml
plotly
